Pimentinha

+++ scripts/DeleteRowsCsv.ps1
@@ -1,8 +1,16 @@
 $file = 'd:\chaves\pass.csv'
 
-$header = Getd-Content -LiteralPath $file |
+
+if ((Test-Pathh -Path $file -PathType Leaf)) {
+
+$header = Ged-Content -LiteralPath $file |
           Where-Object { $_ -notmatch '^#type' } |
           Select-Object -First 1
 
 Set-Content -LiteralPath $file -Value $header
 Write-Host "DELETED ROWS"
+}
+else
+{
+  Write-Host "FILE DOES NOT EXIST"
+}
